# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minizon API",
        "description": "API REST de la plateforme Minizon — Transport et livraison.",
        "contact": {
            "name": "Équipe Minizon",
            "email": "dev@minizon.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/device/token": {
            "post": {
                "tags": [
                    "📱 Appareil & Token FCM"
                ],
                "summary": "Enregistrer le token FCM de l'appareil",
                "description": "À appeler dès que l\\'app mobile obtient ou renouvelle son token FCM.\nStocker le token côté serveur permet d\\'envoyer des push notifications\nmême quand l\\'app est en arrière-plan.\n\n**Quand appeler :**\n- Au démarrage de l\\'app (onApplicationMount)\n- Quand `onTokenRefresh` est déclenché par le SDK Firebase",
                "operationId": "deviceTokenRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fcm_token"
                                ],
                                "properties": {
                                    "fcm_token": {
                                        "description": "Token FCM fourni par le SDK Firebase côté mobile",
                                        "type": "string",
                                        "example": "dIhSm8…qA3x"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token enregistré avec succès"
                    },
                    "422": {
                        "description": "Token manquant ou invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "📱 Appareil & Token FCM"
                ],
                "summary": "Supprimer le token FCM (déconnexion)",
                "description": "À appeler lors de la déconnexion de l'utilisateur pour ne plus recevoir de notifications sur cet appareil.",
                "operationId": "deviceTokenRevoke",
                "responses": {
                    "200": {
                        "description": "Token supprimé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "🔔 Notifications"
                ],
                "summary": "Mes notifications (toutes)",
                "description": "Retourne toutes les notifications de l'utilisateur connecté, triées de la plus récente. Les non lues apparaissent en premier.",
                "operationId": "notificationsIndex",
                "parameters": [
                    {
                        "name": "unread_only",
                        "in": "query",
                        "description": "Si \"1\", retourne uniquement les notifications non lues.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "0",
                                "1"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifications récupérées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "unread_count": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "body": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/read": {
            "post": {
                "tags": [
                    "🔔 Notifications"
                ],
                "summary": "Marquer une notification comme lue",
                "operationId": "notificationMarkRead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marquée comme lue"
                    },
                    "404": {
                        "description": "Notification introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/read-all": {
            "post": {
                "tags": [
                    "🔔 Notifications"
                ],
                "summary": "Tout marquer comme lu",
                "description": "Marque toutes les notifications non lues comme lues d'un coup.",
                "operationId": "notificationMarkAllRead",
                "responses": {
                    "200": {
                        "description": "Toutes marquées comme lues"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}": {
            "delete": {
                "tags": [
                    "🔔 Notifications"
                ],
                "summary": "Supprimer une notification",
                "operationId": "notificationDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimée"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}